Make sourcedeductionservice 2.3 and 2.4 compat
