+++ src/Model/SourceDeductionService.php
@@ -4,59 +4,24 @@ declare(strict_types=1);
 
 namespace Ampersand\DisableStockReservation\Model;
 
-use Magento\Framework\Exception\LocalizedException;
-use Magento\Inventory\Model\SourceItem\Command\DecrementSourceItemQty;
-use Magento\InventoryApi\Api\Data\SourceItemInterface;
-use Magento\InventoryConfigurationApi\Api\Data\StockItemConfigurationInterface;
-use Magento\InventoryConfigurationApi\Api\GetStockItemConfigurationInterface;
-use Magento\InventorySalesApi\Api\GetStockBySalesChannelInterface;
-use Magento\InventorySourceDeductionApi\Model\GetSourceItemBySourceCodeAndSku;
+use Ampersand\DisableStockReservation\Model\SourceDeductionService\SourceDeductionServiceFactory;
 use Magento\InventorySourceDeductionApi\Model\SourceDeductionRequestInterface;
 use Magento\InventorySourceDeductionApi\Model\SourceDeductionServiceInterface;
 
 class SourceDeductionService implements SourceDeductionServiceInterface
 {
     /**
-     * Constant for zero stock quantity value.
+     * @var SourceDeductionServiceFactory
      */
-    private const ZERO_STOCK_QUANTITY = 0.0;
+    private $sourceDeductionServiceFactory;
 
     /**
-     * @var GetSourceItemBySourceCodeAndSku
-     */
-    private $getSourceItemBySourceCodeAndSku;
-
-    /**
-     * @var GetStockItemConfigurationInterface
-     */
-    private $getStockItemConfiguration;
-
-    /**
-     * @var GetStockBySalesChannelInterface
-     */
-    private $getStockBySalesChannel;
-
-    /**
-     * @var DecrementSourceItemQty
-     */
-    private $decrementSourceItem;
-
-    /**
-     * @param GetSourceItemBySourceCodeAndSku $getSourceItemBySourceCodeAndSku
-     * @param GetStockItemConfigurationInterface $getStockItemConfiguration
-     * @param GetStockBySalesChannelInterface $getStockBySalesChannel
-     * @param DecrementSourceItemQty $decrementSourceItem
+     * @param SourceDeductionServiceFactory $sourceDeductionServiceFactory
      */
     public function __construct(
-        GetSourceItemBySourceCodeAndSku $getSourceItemBySourceCodeAndSku,
-        GetStockItemConfigurationInterface $getStockItemConfiguration,
-        GetStockBySalesChannelInterface $getStockBySalesChannel,
-        DecrementSourceItemQty $decrementSourceItem
+        SourceDeductionServiceFactory $sourceDeductionServiceFactory
     ) {
-        $this->getSourceItemBySourceCodeAndSku = $getSourceItemBySourceCodeAndSku;
-        $this->getStockItemConfiguration = $getStockItemConfiguration;
-        $this->getStockBySalesChannel = $getStockBySalesChannel;
-        $this->decrementSourceItem = $decrementSourceItem;
+        $this->sourceDeductionServiceFactory = $sourceDeductionServiceFactory;
     }
 
     /**
@@ -64,84 +29,6 @@ class SourceDeductionService implements SourceDeductionServiceInterface
      */
     public function execute(SourceDeductionRequestInterface $sourceDeductionRequest): void
     {
-        $sourceCode = $sourceDeductionRequest->getSourceCode();
-        $salesChannel = $sourceDeductionRequest->getSalesChannel();
-        $stockId = $this->getStockBySalesChannel->execute($salesChannel)->getStockId();
-        $sourceItemDecrementData = [];
-        foreach ($sourceDeductionRequest->getItems() as $item) {
-            $itemSku = $item->getSku();
-            $qty = $item->getQty();
-            $stockItemConfiguration = $this->getStockItemConfiguration->execute(
-                $itemSku,
-                $stockId
-            );
-
-            if (!$stockItemConfiguration->isManageStock()) {
-                //We don't need to Manage Stock
-                continue;
-            }
-
-            $sourceItem = $this->getSourceItemBySourceCodeAndSku->execute($sourceCode, $itemSku);
-
-            /*
-             * Fix a problem when a product has a large negative quantity, and an order with that product is canceled
-             * from backend.
-            */
-            $salesEvent = $sourceDeductionRequest->getSalesEvent();
-            if ($salesEvent->getType() ==
-                \Magento\InventorySalesApi\Api\Data\SalesEventInterface::EVENT_ORDER_CANCELED) {
-                if (($sourceItem->getQuantity() - $qty) < 0) {
-                    $sourceItem->setQuantity($sourceItem->getQuantity() - $qty);
-                    $stockStatus = $this->getSourceStockStatus(
-                        $stockItemConfiguration,
-                        $sourceItem
-                    );
-                    $sourceItem->setStatus($stockStatus);
-                    continue;
-                }
-            }
-
-            if (($sourceItem->getQuantity() - $qty) >= 0) {
-                $sourceItem->setQuantity($sourceItem->getQuantity() - $qty);
-                $stockStatus = $this->getSourceStockStatus(
-                    $stockItemConfiguration,
-                    $sourceItem
-                );
-                $sourceItem->setStatus($stockStatus);
-                $sourceItemDecrementData[] = [
-                    'source_item' => $sourceItem,
-                    'qty_to_decrement' => $qty
-                ];
-            } else {
-                throw new LocalizedException(
-                    __('Not all of your products are available in the requested quantity.')
-                );
-            }
-        }
-
-        if (!empty($sourceItemDecrementData)) {
-            $this->decrementSourceItem->execute($sourceItemDecrementData);
-        }
-    }
-
-    /**
-     * Get source item stock status after quantity deduction.
-     *
-     * We revert https://github.com/magento/inventory/commit/6f20ba6f96c1d124f8d0770745261a90c907f9e4.
-     *
-     * @param StockItemConfigurationInterface $stockItemConfiguration
-     * @param SourceItemInterface $sourceItem
-     *
-     * @return int
-     */
-    private function getSourceStockStatus(
-        StockItemConfigurationInterface $stockItemConfiguration,
-        SourceItemInterface $sourceItem
-    ): int {
-        $sourceItemQty = $sourceItem->getQuantity() ?: self::ZERO_STOCK_QUANTITY;
-
-        return $sourceItemQty === $stockItemConfiguration->getMinQty() && !$stockItemConfiguration->getBackorders()
-            ? SourceItemInterface::STATUS_OUT_OF_STOCK
-            : SourceItemInterface::STATUS_IN_STOCK;
+        $this->sourceDeductionServiceFactory->create()->execute($sourceDeductionRequest);
     }
 }
